requirements.txt
# API Framework
fastapi==0.109.0
uvicorn==0.27.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# LLM Integration
anthropic==0.18.1
openai==1.12.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Utilities
python-dotenv==1.0.0
pydantic-settings==2.1.0
